+++ src/main/java/com/stratio/qa/specs/CommonG.java
@@ -1355,6 +1355,17 @@ public class CommonG {
         return seleniumCookies;
     }
 
+    public boolean cookieExists(String cookieName) {
+        if (this.getSeleniumCookies() != null && this.getSeleniumCookies().size() != 0) {
+            for (org.openqa.selenium.Cookie cookie: this.getSeleniumCookies()) {
+                if (cookie.getName().contains(cookieName)) {
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+
     public void setSeleniumCookies(Set<org.openqa.selenium.Cookie> cookies) {
         this.seleniumCookies = cookies;
     }